extracted OneTimePasswordGenerator functionality and added tests
